refactor: proxy/frontend api in feature oriented architecture
